+++ axelor-production/src/main/java/com/axelor/apps/production/service/MrpServiceProductionImpl.java
@@ -420,7 +420,7 @@ public class MrpServiceProductionImpl extends MrpServiceImpl {
           this.assignProductLevel(subBillOfMaterial, level);
 
           if (subProduct.getDefaultBillOfMaterial() != null) {
-            this.productMap.put(subProduct.getId(), this.getMaxLevel(subProduct, level));
+            this.assignProductLevel(subProduct.getDefaultBillOfMaterial(), level);
           }
         }
       }